Внес исправления

+++ migration/src/main/resources/db/changelog/v01/courier.sql
@@ -2,7 +2,7 @@ create sequence if not exists courier_seq;
 
 create table if not exists courier
 (
-    courier_id integer not null default nextval ('courier_seq'),
+    courier_id bigint not null default nextval ('courier_seq'),
     phone varchar(11) not null,
     coordinates point not null,
     status varchar(255) not null,
