+++ migration/src/main/resources/db/changelog/v01/restaurant.sql
@@ -2,7 +2,7 @@ create sequence if not exists restaurant_seq;
 
 create table if not exists restaurant
 (
-    restaurant_id integer not null default nextval ('restaurant_seq'),
+    restaurant_id bigint not null default nextval ('restaurant_seq'),
     address varchar(256) not null,
     status varchar(128),
     constraint restaurant_pk primary key (restaurant_id)
